refactor: separate r2dbc converter classes
